ExpenseDocAI/src/requirements.txt
# Core dependencies
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-health-check==3.17.0
django-redis==5.4.0
whitenoise==6.6.0

# Database
mysqlclient==2.2.3
redis==5.0.1

# Image processing
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
numpy==1.26.4

# AI/ML
torch==2.2.0
transformers==4.37.2
layoutlm==0.1.0

# Storage
django-storages==1.14.2
boto3==1.34.34

# Security
django-axes==6.1.1
django-csp==3.7
django-ratelimit==4.1.0

# WSGI/FCGI
flup6==1.1.1
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0 